requirements.txt
asgiref==3.8.1
autopep8==2.3.1
click==8.1.7
Django==5.0.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-js-asset==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
gunicorn==23.0.0
lxml==5.2.2
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.10
pycodestyle==2.12.0
python-docx==1.1.2
python-dotenv==1.0.1
setuptools==75.3.0
sqlparse==0.5.0
svgwrite==1.4.3
tomli==2.0.1
Tree==0.2.4
typing_extensions==4.12.2
whitenoise==6.8.2
